+++ backend/src/main/java/com/bootcamp/learning/bootcamp/repository/AccountsRepository.java
@@ -3,6 +3,6 @@ package com.bootcamp.learning.bootcamp.repository;
 import com.bootcamp.learning.bootcamp.entity.Accounts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountsRepository extends JpaRepository<Accounts, Integer> {
+public interface AccountsRepository extends JpaRepository<Accounts, Long> {
     // custom queries
 }
